feat: have routing by default for new applications
