+++ src/ng-new/application/_files/__sourcedir__/home/home.component.html
@@ -1,0 +1,6 @@
+<!--The content below is only a placeholder and can be replaced.-->
+<StackLayout<% if (theme) { %> class="p-20"<% } %>>
+  <Label text="Tap the button"<% if (theme) { %> class="h1 text-center"<% } %>></Label>
+  <Button text="tap" (tap)="onTap()"<% if (theme) { %> class="btn btn-primary btn-active"<% } %>></Button>
+  <Label [text]="getMessage()"<% if (theme) { %> class="h2 text-center"<% } %> textWrap="true"></Label>
+</StackLayout>
