+++ src/ng-new/application/_files/__sourcedir__/home/home.component.ts
@@ -1,0 +1,26 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: '<%= prefix %>-home',
+  templateUrl: './home.component.html',
+  styleUrls: ['./home.component.css'],
+  moduleId: module.id,
+})
+export class HomeComponent implements OnInit {
+  private counter = 42;
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+  public getMessage() {
+    return this.counter > 0 ?
+      `${this.counter} taps left` :
+      'Hoorraaay! You unlocked the NativeScript clicker achievement!';
+  }
+
+  public onTap() {
+    this.counter--;
+  }
+}
