--- src/ng-new/application/_routing-files/__sourcedir__/app-routing.module.ts
@@ -1,18 +1,0 @@
-import { NgModule } from '@angular/core';
-import { NativeScriptRouterModule } from 'nativescript-angular/router';
-import { Routes } from '@angular/router';
-
-import { ItemsComponent } from './item/items.component';
-import { ItemDetailComponent } from './item/item-detail.component';
-
-const routes: Routes = [
-    { path: '', redirectTo: '/items', pathMatch: 'full' },
-    { path: 'items', component: ItemsComponent },
-    { path: 'item/:id', component: ItemDetailComponent },
-];
-
-@NgModule({
-    imports: [NativeScriptRouterModule.forRoot(routes)],
-    exports: [NativeScriptRouterModule]
-})
-export class AppRoutingModule { }
