--- src/ng-new/application/_routing-files/__sourcedir__/item/item-detail.component.html
@@ -1,8 +1,0 @@
-<ActionBar title="Details"<% if (theme) { %> class="action-bar"<% } %>></ActionBar>
-<FlexboxLayout flexDirection="column" class="page">
-    <FlexboxLayout class="m-15">
-        <Label<% if (theme) { %> class="h2"<% } %> [text]="item.id + '. '"></Label>
-        <Label<% if (theme) { %> class="h2"<% } %> [text]="item.name"></Label>
-    </FlexboxLayout>
-    <Label<% if (theme) { %> class="h4"<% } %> [text]="item.role"></Label>
-</FlexboxLayout>
