--- src/ng-new/application/_routing-files/__sourcedir__/item/item-detail.component.ts
@@ -1,24 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-
-import { Item } from './item';
-import { ItemService } from './item.service';
-
-@Component({
-    selector: '<%= prefix %>-details',
-    moduleId: module.id,
-    templateUrl: './item-detail.component.html',
-})
-export class ItemDetailComponent implements OnInit {
-    item: Item;
-
-    constructor(
-        private itemService: ItemService,
-        private route: ActivatedRoute
-    ) { }
-
-    ngOnInit(): void {
-        const id = +this.route.snapshot.params['id'];
-        this.item = this.itemService.getItem(id);
-    }
-}
