--- src/ng-new/application/_routing-files/__sourcedir__/item/item.ts
@@ -1,5 +1,0 @@
-export interface Item {
-    id: number;
-    name: string;
-    role: string;
-}
