--- src/ng-new/application/_routing-files/__sourcedir__/item/items.component.html
@@ -1,8 +1,0 @@
-<ListView [items]="items" class="list-group">
-    <ng-template let-item="item">
-        <Label [nsRouterLink]="['/item', item.id]"
-            [text]="item.name"<% if (theme) { %>
-            class="list-group-item"<% } %>
-        ></Label>
-    </ng-template>
-</ListView>
